Revisados permisos de todos los controladores de creación y edición para que se tengan en cuenta a la hora de acceder a las páginas o secciones correspondientes directamente a través de la barra de direcciones del navegador. Cambios en la interfaz que lista los cuadros de mando de una unidad. Otras correcciones...

+++ icasus/app_code/modules/cuadros/cuadro_borrar.php
@@ -9,27 +9,36 @@
 // Descripcion: Borra un cuadro de mando
 //---------------------------------------------------------------------------------------------------
 
-global $smarty;
-global $_usuario;
+global $usuario;
 
 if (filter_has_var(INPUT_GET, 'id_cuadro'))
 {
     $id_cuadro = filter_input(INPUT_GET, 'id_cuadro', FILTER_SANITIZE_NUMBER_INT);
-    $id_usuario = $_usuario->id_usuario;
     $cuadro = new Cuadro();
     $cuadro->load("id = $id_cuadro");
-    $panel = new Panel();
-    $paneles_cuadro = $panel->Find("id_cuadro=$id_cuadro");
-    if (!$paneles_cuadro)
+
+    //Comprobamos los permisos del usuario
+    if ($control || $usuario->id == $cuadro->id_usuario)
     {
-        $exito = MSG_CUADRO_MANDO_BORRADO . " " . $cuadro->nombre;
-        $cuadro->delete();
-        header("Location: index.php?page=cuadro_listar&id_entidad=$cuadro->id_entidad&exito=$exito");
+        $panel = new Panel();
+        $paneles_cuadro = $panel->Find("id_cuadro=$id_cuadro");
+        if (!$paneles_cuadro)
+        {
+
+            $exito = MSG_CUADRO_MANDO_BORRADO . " " . $cuadro->nombre;
+            $cuadro->delete();
+            header("Location: index.php?page=cuadro_listar&id_entidad=$cuadro->id_entidad&exito=$exito");
+        }
+        else
+        {
+            $aviso = MSG_CUADRO_MANDO_BORRAR;
+            header("Location: index.php?page=cuadro_listar&id_entidad=$cuadro->id_entidad&aviso=$aviso");
+        }
     }
     else
     {
-        $aviso = MSG_CUADRO_MANDO_BORRAR;
-        header("Location: index.php?page=cuadro_listar&id_entidad=$cuadro->id_entidad&aviso=$aviso");
+        $error = ERR_PERMISOS;
+        header("Location: index.php?page=cuadro_listar&id_entidad=$cuadro->id_entidad&error=$error");
     }
 }
 else
